ScriptableObjectTestingV0.4
Everything is broken, life is hell, BUT i fixed the issue with scriptable objects resetting on scene change by making the other classes also be scriptable object assets. UNFORTUNATELY that didn't fix it entirely, and now it resets when unity is reopened instead.
Dread it, run from it, scriptable objects reset all the same.